Add service for item tokens

+++ backend/src/main/java/hu/schbme/paybasz/station/model/ItemTokenEntity.java
@@ -15,13 +15,13 @@ public class ItemTokenEntity {
 	@Id
 	@Column
 	@GeneratedValue
-	private Long id;
+	private Integer id;
 
 	@Column(nullable = false, name = "user_id")
-	private Long userId; // no Hibernate magic needed
+	private Integer userId; // no Hibernate magic needed
 
 	@Column(nullable = false, name = "item_id")
-	private Long itemId; // no Hibernate magic needed
+	private Integer itemId; // no Hibernate magic needed
 
 	@Column(nullable = false)
 	private Integer count;
